requirements.txt
typer==0.6.1
GitPython==3.1.27